refactor: move route files to routes folder
