+++ src/app.module.ts
@@ -1,11 +1,11 @@
 import { Module } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { UserModule } from './user/user.module';
-import { TrackModule } from './track/track.module';
-import { ArtistModule } from './artist/artist.module';
-import { AlbumModule } from './album/album.module';
-import { FavouritesModule } from './favourites/favourites.module';
+import { UserModule } from './routes/user/user.module';
+import { TrackModule } from './routes/track/track.module';
+import { ArtistModule } from './routes/artist/artist.module';
+import { AlbumModule } from './routes/album/album.module';
+import { FavouritesModule } from './routes/favourites/favourites.module';
 
 @Module({
   imports: [
