+++ src/routes/album/entities/album.entity.ts
