+++ src/routes/artist/dto/create-artist.dto.ts
