+++ src/routes/artist/dto/update-artist.dto.ts
