+++ src/routes/artist/entities/artist.entity.ts
